requirements.txt
pygame-ce>=2.5.0
python-chess>=1.10.0
torch>=2.0.0
numpy>=1.24.0
pybind11>=2.11.0
